src/main/java/com/aryzhkov/movieland/dao/jdbc/mapper/CountryRowMapper.java
package com.aryzhkov.movieland.dao.jdbc.mapper;

import com.aryzhkov.movieland.entity.Country;
import org.springframework.jdbc.core.RowMapper;

import java.sql.ResultSet;
import java.sql.SQLException;

public class CountryRowMapper implements RowMapper<Country> {

    @Override
    public Country mapRow(ResultSet resultSet, int i) throws SQLException {
        Country country = new Country();

        country.setId(resultSet.getInt("COUNTRY_ID"));
        country.setName(resultSet.getString("COUNTRY_NAME"));

        return country;
    }
}
